Borra clase TemaTo de la jerarquía porque no es necesaria
